Improve the factoring of the stylesheet used to produce the new format Struts documentation.  Add a stylesheet that takes the same source code and produces a TLD file.
Add a new-format page for a set of tags focused on creating new beans
(available to the rest of the page) from a variety of sources.  The tags
themselves do not exist yet, but they will shortly.

The new-format documentation will shortly be posted on the Jakarta web
site page for Struts, at <http://jakarta.apache.org/struts>.  Be sure to
check it out and make suggestions for improving the usability of these
documents.