During a "forget" operation:
If the Xid is for the wrong transaction, throw an XAException immediately.
Otherwise, send the transaction failed notification.
Don't throw any other XAExceptions until after that so the notification
will go through and the connection will be removed from the pool.

The XAER_PROTO error probably shouldn't be thrown here at all, in the long
run, but we'll take this one step at a time.