Implement changes suggested by myself and Jay Walters: When fetching data into a field, use the field's class to determine how to get the data from the ResultSet.  Avoids problems with "getObject" and DB-specific types. With the class of the field, we know better than the DB drivers what JDBC type we want.
If we can implement this for writing to PreparedStatements as well, then
we can trim a lot of fat out of the XML configuration files (we will only
need to know SQL types for creating DB tables).

Also, convert java.util.Date object to java.sql.Date, Time, or Timestamp
as appropriate before setting.