Server adapter refactoring.
The new code makes the implementation closer to the design patterns
we chosed ( i.e. Interceptors act as an adapter between an external API - the
protocol implementation - and tomcat internal representation ).

This is the first step in documenting Ajp13 - the protocol is now in
a single file, easier to read and enhance. We no longer try
to make ajp13 a "generic-protocol" - we'll stay with TCP-specific. JNI
doors or other protocols have different requirement and we can
deal with that after ajp13 is stable.

All other files in this package will slowly disapear, and unless you find
a design problem with the new organization I'll remove them.
As usually that will happen after the new code is tested and stable.
( now it compile, but I haven't even run it )

TODO:
- make sure it runs
- document Ajp13.
- change Ajp13, Jni and Http to use MessageBytes and make sure they are
optimized ( we should be now in the race for the last 10-20 object allocations
per request)
- enhance Ajp13 with all 2-directional calls we need
- !!! make sure the buffers are properly used ( one important optimization
will be to lazy-decode the buffers, but that means the buffer can't
be reused for other callbacks. The current scheme works fine, and
I don't expect this to be a problem ). Since this is a low-level
protocol with a clear interface and function it's important to tune it
very well.