- Core is now independent of facades, i.e. we have a real layering and module separation !!!
- Added deprecate comments to FacadeManager - we know there are better
ways to get to the real object from trusted webapps, and the 1-1 relation
between Request-RequestFacade, etc have serious security problems.

- start to use the recyclable ServerCookie. We still need to add code
like in MimeHeaders to enable recycling. The best solution (IMHO) is to
transform MimeHeaders in a general-purpose MultiMap ( string -> string[],
with most elements having only one value ) and start using it not only
for headers, but for cookies and parameters too. Right now normal header
processing can be done with 0 GC ( constant memory ), we should extend
 that to all other request components.