Fix OutputBuffer bug ( at least one :-)
The problem was simple - recycle() didn't reset the buffer of the
writer ( convertor ), and in some cases ( for example if the connection
is broken ) the buffer was not flushed.

XXX Need to find out why the buffer wasn't flushed - I think we
should assert( ob.flush() called after each request ).

PR:
Obtained from:
Submitted by:
Reviewed by: