As proposed, removed ContextInterceptor and RequestInterceptor.
In order to write a tomcat module you need to extend BaseInterceptor. With
the new hook mechanism ( and thanks to introspection ) there is no
problem with the number of interceptors and no need to have 2 different
types.

This is not only simpler, but more efficient too ( not very big difference,
but for the record).

I don't think any external ( i.e. used to integrate tomcat) interceptor
is affected, any normal one should already extend BaseInterceptor.

We now need to clean up the methods and code dealing with interceptors
( 1/2 ), and maybe later on add ContextEvent, ServerEvent, etc for
async events - I have them in my workspace but I don't think it's a good
idea for 3.3...

There is still a big difference between hooks in the critical path
and async ( context ) hooks - we can afford to generate new Event
objects and throw exceptions for the last ones.