As proposed (long time ago), moved tomcat.logging to tomcat.util.log
Since this is just code move, with no benefit other than esthetics,
I also optimized a bit the logger ( no more StringBuffer.toString,
reuse the StringBuffer, etc).

Logger.Helper is now renamed to Log() ( this is the main class used by
clients, it deserves a simple name ).

TomcatLogger is now renamed QueueLogger - it has nothing to do with tomcat,
but it put all log messages in a queue to save time in the critical
path ( we still have a sync. - but I don't think that can be done without
sync, at least not in 2 hours )

I also moved some code arround to simplify a bit QueueLogger.