Moved doPreServletInit/doPostServletInit and Destroy from contextmanager to ServletWrapper.
There is no benefit in having it in ContextManager ( and same will happen
with most doXXX methods), while having them in the code that actually
needs it seems better.
( for example it's easier to understand when the hooks are called
and what's the return convention - i.e. errors are ignored )

It seems pre/postServletDestroy wasn't called when it should.