Few big changes in ContextManager.
- removed RequestInterceptors and ContextInterceptors. We only have
Interceptors ( no need to make distinction )

- most of the doHOOK moved next to the caller, no need to
call context manager to intermediate. That will also help when/if we
add some mechanism to define dynamic hooks.

- IMPORTANT - documented the order of startup ( at least the way I understand
it), and make sure no hooks are called until CM is initialized.
We still need to make sure that
1. No request is served until CM is in STARTED state
2. A context will not serve requests until it's in ACTIVE state.

- A lot of cleanup and simplifactions resulting from this ( the state of
ContextManager should be consistent now). For example we can move more
functionality in interceptors ( fixing the paths is now done in
DefaultCMSetter and it's much improved, etc).

So far I thought MessageBytes and OutputBuffer are the most important changes
for 3.3. Now I think making sure we have a clear definition of state
for CM and interceptors is much more important. Please help and review
( at least the comments at the beginning of ContextManager )