Initial commit for the webdav module from catalina.
Most of the junk is removed, but a bit more refactoring is still needed -
the cache generates a lot of garbage and the number of threads need to
be reduced ( and integrated with the server's thread management )

The new tomcat3 module has no external dependency ( except servlet22.jar ).
It doesn't work yet, but compiles fine ( it's very close to working :-)

Probably the resources should go to tomcat.util.resources, as we may reuse
them ( or at least the cache ) for static files ( well, that's just for
benchmark purpose, I don't think any decent admin would turn this on,
using a large cache in java will kill everything else - of course,
if only one file is cached, like in a "ab" benchmark, it looks ok ... )

The only missing part is the ServerLiaison, that will allow it to plug
into any servlet container.