A small round of optimizations - removed few dozens of un-needed allocations and GC.
Now we are down to 11 objects per simple request - still not zero, but close
enough. I think that concludes the memory tunning for the "common path"
( i.e. the overhead common to all requests ).

We can now start doing "real" performance in 3.3 !!! ( from my point
of view, nothing so far was more than removing garbage )

( of course, Cookies and Parameters have to be rewritten after the MimeHeaders
model, plus a refactoring of MimeHeaders )

( BTW, all 11 objects are allocated in interceptors, no change in core
will be needed to remove them - one is the Socket and it's not easy to
reuse that - so probably we'll have 1 object per request instead of 0.
 )

Details:

- Continue the work on Request - now "serverName" ( the virtual host ) is
a reusable MessageByte. There are only few Strings left, most of them not
used in most cases.

- small optimization in RequestUtil - no need to do the loop if the string
is not encoded ( that saves few Strings/request + a loop )

- small optimization in SessionUtil - the string allocation is not needed

- small optimization in AccessInterceptor - avoid string allocation
( it'll be further optimized after we're done with the core, I'm working
on few better tools for string manipulation )

- small optimization in SimpleMapper1 - use/pass MessageBytes instead of
converting to String.

- same in PrefixMapper, use unsynchronized Hashtable ( SimpleHashtable - from
crimson :-), pass strings

- added getChars in MessageBytes

PR:
Obtained from:
Submitted by:
Reviewed by: