Fix the build for JDK1.1
Since it's expected that the optimizations will use more JDK1.2 features
( in optional modules, of course :-) we want to make sure that tomcat's
core remains usable on JDK1.1 and the optimized code is indeed optional.

The rule is simple - no JDK1.2 specific code in core, in the basic modules
or in utils used by basic modules ( if you do, the nightly build should
catch it ).

Use
<exclude name="org/apache../MyJdk12Class.java" unless="jdk12.present" />
for all the 1.2 modules ( and make sure they are not enabled by default
in server.xml )

XXX need an exclude/unless for server.xml


PR:
Obtained from:
Submitted by:
Reviewed by: