Start work on Cookies, Parameters, Headers.
The code in Cookies is almost ready, all cookie processing was re-done
from scratch - and will support the full spec, not only name/values.
It'll also be much faster, and the object model is better ( no need
for artificial "helpers").

Parameter parsing is also rewritten ( and more efficient ).

Both Parameters and Cookies are now (almost) GC-free. The code will be more
eficient if there are only few cookies/parameters (2..10 ?), for requests
with many params we'll add the optimizations later.

The code is not used right now - it'll gradually move in.