Added a new module, based on refactoring part of DefaultCMSetter.
This module will "fix" any relative path used in tomcat's configs,
using the same rules as before. The main difference is that all the "fixing" is
grouped in a single replaceable module - that should make the code
easier to read and fix.

The module will replace DefaultCMSetter after some more testing. It needs more
documentation of the rules ( based on what we think is the right way to
do that, not what happened to happen before )