Port exception propagation behavior from Tomcat 3.2. Exceptions that occur in included servlets are throw back to their callers.  The exception is handled only if it occurs or when it reaches the top level handler.
The implementation differs from that of Tomcat 3.2. Where Tomcat 3.2
catches and rethrows the exception at each level, this implementation tries
to catch and save the exception.  I think this makes the flow of execution
easier to control.  We can give interceptors a chance to examine the
exceptions.  Now, only RequestDispatcherImpl rethrows the exceptions.

There are several places that exceptions are saved. Init exceptions and
UnavailableExceptions thrown in the service() method are saved in Handler
along with a "permanent" indication. postServletInit interceptors can find it
here.  Also, postServletInit is called even if the init() method throws an
exception.

The first exception that occurs during request processing is saved on the
Response along with the "include" URI, if present.  This exception will not
be overwritten if subsequent exceptions occur.

Exceptions during request processing are also saved on the Request.  But
unlike the Response exception, the Request exception is overwritten with
each new exception that occurs.

In RequestDispatcherImpl, setting of the "included" flag was restored in
includeNamed() and forwardNamed() since it is used to control error handling
behavior.

JspInterceptor has been updated to set an initial exception if translation or
compilation fails.  This way these error won't always show up as a 404 error.
Also, setting the compileTime from the java file was removed.  After a restart,
this prevented the error from being recreated.  Thus, a new java file will be
created after each restart for a non-compiling JSP, but this is better than
creating one on each request.

Watchdog still passes, but these changes need more review and testing.
Handling of destroy() wasn't addressed in these changes.  This needs
review in the context of these changes.