Major rewrite of JspInterceptor. IMHO what used to be ServletWrapper and all related classes are now in a much better shape. That should be the last big change - we should start building the first milestone for 3.3 soon. ( I want to do yet-another review of the core and remove everything that is not used or can be simplified before that )
The new code removes few optimizations ( for simplicity ) - most of them
were part of the initial loading or compilation of JSPs, where the
compiler is taking most of the time.

There is yet another simplification we can do ( separate the ServletHandler
and add a JspHanlder ), but so far ServletHandler doesn't seem too complex
and handling jsps is a very small overhead ( far away from what it used to be).