Synchronize setting of data for UnavailableException thrown from service(). An UnavailableException returning a getUnavailableSeconds() = 0 could be expired out from under the request trying to set it if another request came in with just the right timing.
We may want to set a minimum unavailable time.  If we keep setting and
expiring an UnavailableException that returns getUnavailableSeconds() = 0,
we could end up synchronizing a lot.

A few additional corrections to get UnavailableException working as
expected.

Error handling still needs updating so that UnavailableExceptions
return a status code of 503 instead of 500.