Some simplification.
Removed saveError method from Handler.  setErrorException() and
setErrorURI() are called directly where needed.  setErrorURI() should be
called with a non-null error URI only if the URI causing the error differs from
the requestURI found on the associated request.  Thus, this needs to
happen only in RequestDispatcherImpl.

Added handling to report UnavailableExceptions as 503 errors.