Explicit imports in core. There are few more dependencies that should be removed ( and moved to the facade, as high-level constructs - Principal, etc ).
Next step is to do the same thing in util, and remove ( move out ) all
utils that are no longer used.

The place for "general utils" is in a "jpr" ( i.e. java-APR )-like project,
and even the curently-used utils shouldn't be part of tomcat ( in the
same way as Apache2.0 makes a clear difference and uses 2 CVS trees - one
for the general-code and one for the library )