Removed few bad tests (IMHO) that tested for tomcat-specific behavior ( that was wrong ).
The problem is that in previous versions of tomcat, errors RequestDispatcher.include
were not propagated. Thanks to Larry and other's work on error handling we do now the
 right thing.

One of the tests was also assuming the buffer is not reset before the error message is
shown - I think it's very important to do reset the buffer.

Also, in IncludeMismatch1 the test assumed the error handler will be called - but the
spec requires to forward the exception, and it is dealt with in the calling servlet,
so no error page will be displayed.

Another bad test was Set-Cookie2 - which belongs to the wrong Cookie spec.