Add transactional token support to Struts, and use it in the example application to ensure that a user does not go to the edit registration screen, press Save, use the back arrow on their browser, and press Save again.
Use of transactional token support is totally under the control of your
Action classes.  See the following changes in the example app:
* EditRegistrationAction - Call saveToken() just before forwarding
* SaveRegistrationAction - Call isTokenValid() before accepting this
  transaction, then call resetToken() to make sure the same transaction
  is not accepted twice.

The approach taken has been inspired by the input of many Struts users,
including Robert Leland, William Jaynes, and David Geary.  In addition,
a similar model is used in:

	Fields, Duane K. and Kolb, Mark A., WEB DEVELOPMENT WITH
		JAVA SERVER PAGES, Manning Publications, 2000

NOTE:  There is still a problem with the subscription.jsp page, probably
related to recent changes in BeanUtils and PropertyUtils, that I am
going to look at next.