Added missing file, needed to run the self-test ( was moved from share/ since was not a required part of tomcat, and is used only in the test app ).
There are many enhancements needed to make it really easy to run the
tests. I guess we'll do them as we go.

I also did a small change - GTest new extends Task and uses log() instead
of System.out(). ( I'll check in a jsp file that runs the test - the change is
needed to send the output to the right destination ).

Note: this is a change to the CVS HEAD. IMHO there is nothing to stop
fixes to go in - as long as we don't call it anyway. I'll commit a number
of fixes over the weekend - including restoring the JDK1.1 build ( plus
a solution that will simplify supporting JDK1.1 ), and I'll probably finish
the .jsp taglib that runs the tests. Lazy consensus should work as before -
if anyone has a problem with any checkin, please -1 it.