Enhancements ( and cleanup ) of GTest.java.
This is a general-purpose tool ( HttpClient and the Matcher can be
used as beans in many places ). Other enhancements will follow.

The goal is to make it easier to run and add tests, and allow more complex
test cases to be deveolped. Ramesh's changes in watchdog should also be
incorporated.

Historic Note: I wrote the first version of GTest based on previous
tester that was part of tomcat's sanity-test. GTest ( or similar tools)
is used in other projects and ( AFAIK ). It is a nice dream to have a
common jakarta repository for all general-purpose tools and utilities.

The GTest is far for beeing complete ( the current code compiles, but
hasn't been tested ). I'll send a lot of information after I finish
with the code - including ( initial ) docs about how to add tests, how
to run the testsuite, etc.

Testing tomcat ( running and developing tests ) will be _fun_.