Second round of changes in the sanity-test running.
You should be able to run the tests by starting tomcat and accessing
http://localhost:8080/admin/test/test.jsp

The old GTest is deprecated, it has been split into HttpClient and
DefaultMatcher ( plus a wrapper that allows 'old-style' calling ).

The goal is to make it easy to run the tests, and to make it easy and
clean to add tests.

I'll probably add a page that allow to run watchdog from the web, but
right now it's more important to get "writing new tests" and
adding regression tests for the bugs we fix as easy and simple as possible.

( long term we should use a dedicated test program - ant is getting
very hard to use embeded without enough API control ).

Changes:
- added a form to test.jsp, allow to run a subset of the tests
- fixes to deal with IntrospectionHelper
- various fixes in the util.test
- more debugging
- dispay the <a> on GET tests, so we can re-run individual tests
manually
- added a sample test using the cleaner ( IMHO ) syntax ( i.e. httpClient
followed by a matcher )

Missing:
- code documentation
- documentation on how to write tests
- fixes, etc.

( it's a test application, should make things simpler - if it doesn't we
should go back to the old one )